+++ packages/cli/test/outputs/clients/index.ts
@@ -1,0 +1,119 @@
+export * from "./account/AccountClient";
+
+export * from "./account/RegisterAccountRequest";
+
+export * from "./account/RegisterAccountRequestCommand";
+
+export * from "./account/RegisterAccountRequestValidator";
+
+export * from "./account/RegisterAccountResponse";
+
+export * from "./account/RegisterAccountResponseValidator";
+
+export * from "./auth/AccessTokenRequest";
+
+export * from "./auth/AccessTokenRequestCommand";
+
+export * from "./auth/AccessTokenRequestValidator";
+
+export * from "./auth/AccessTokenResponse";
+
+export * from "./auth/AccessTokenResponseValidator";
+
+export * from "./auth/AuthClient";
+
+export * from "./auth/RefreshTokenRequest";
+
+export * from "./auth/RefreshTokenRequestCommand";
+
+export * from "./auth/RefreshTokenRequestValidator";
+
+export * from "./auth/RefreshTokenResponse";
+
+export * from "./auth/RefreshTokenResponseValidator";
+
+export * from "./shared/ConflictErrorResponse";
+
+export * from "./shared/ForbiddenErrorResponse";
+
+export * from "./shared/InternalServerErrorResponse";
+
+export * from "./shared/NotFoundErrorResponse";
+
+export * from "./shared/TodoNotChangeableErrorResponse";
+
+export * from "./shared/TodoNotFoundErrorResponse";
+
+export * from "./shared/TodoStatusTransitionInvalidErrorResponse";
+
+export * from "./shared/TooManyRequestsErrorResponse";
+
+export * from "./shared/UnauthorizedErrorResponse";
+
+export * from "./shared/UnprocessableEntityErrorResponse";
+
+export * from "./shared/UnsupportedMediaTypeErrorResponse";
+
+export * from "./shared/ValidationErrorResponse";
+
+export * from "./todo/CreateTodoRequest";
+
+export * from "./todo/CreateTodoRequestCommand";
+
+export * from "./todo/CreateTodoRequestValidator";
+
+export * from "./todo/CreateTodoResponse";
+
+export * from "./todo/CreateTodoResponseValidator";
+
+export * from "./todo/DeleteTodoRequest";
+
+export * from "./todo/DeleteTodoRequestCommand";
+
+export * from "./todo/DeleteTodoRequestValidator";
+
+export * from "./todo/DeleteTodoResponse";
+
+export * from "./todo/DeleteTodoResponseValidator";
+
+export * from "./todo/GetTodoRequest";
+
+export * from "./todo/GetTodoRequestCommand";
+
+export * from "./todo/GetTodoRequestValidator";
+
+export * from "./todo/GetTodoResponse";
+
+export * from "./todo/GetTodoResponseValidator";
+
+export * from "./todo/ListTodosRequest";
+
+export * from "./todo/ListTodosRequestCommand";
+
+export * from "./todo/ListTodosRequestValidator";
+
+export * from "./todo/ListTodosResponse";
+
+export * from "./todo/ListTodosResponseValidator";
+
+export * from "./todo/TodoClient";
+
+export * from "./todo/UpdateTodoRequest";
+
+export * from "./todo/UpdateTodoRequestCommand";
+
+export * from "./todo/UpdateTodoRequestValidator";
+
+export * from "./todo/UpdateTodoResponse";
+
+export * from "./todo/UpdateTodoResponseValidator";
+
+export * from "./todo/UpdateTodoStatusRequest";
+
+export * from "./todo/UpdateTodoStatusRequestCommand";
+
+export * from "./todo/UpdateTodoStatusRequestValidator";
+
+export * from "./todo/UpdateTodoStatusResponse";
+
+export * from "./todo/UpdateTodoStatusResponseValidator";