+++ packages/cli/test/outputs/hono/shared/TodoNotFoundErrorResponse.ts
@@ -1,0 +1,41 @@
+import { HttpResponse, HttpStatusCode } from "@rexeus/typeweaver-core";
+
+export type ITodoNotFoundErrorResponseHeader = {
+  "Content-Type": "application/json";
+};
+
+export type ITodoNotFoundErrorResponseBody = {
+  message: "Todo not found";
+  code: "TODO_NOT_FOUND_ERROR";
+  actualValues: {
+    todoId: string;
+  };
+};
+
+export type ITodoNotFoundErrorResponse = {
+  statusCode: HttpStatusCode.NOT_FOUND;
+  header: ITodoNotFoundErrorResponseHeader;
+  body: ITodoNotFoundErrorResponseBody;
+};
+
+export class TodoNotFoundErrorResponse
+  extends HttpResponse<
+    ITodoNotFoundErrorResponseHeader,
+    ITodoNotFoundErrorResponseBody
+  >
+  implements ITodoNotFoundErrorResponse
+{
+  public override readonly statusCode: HttpStatusCode.NOT_FOUND;
+
+  public constructor(response: ITodoNotFoundErrorResponse) {
+    super(response.statusCode, response.header, response.body);
+
+    if (response.statusCode !== HttpStatusCode.NOT_FOUND) {
+      throw new Error(
+        `Invalid status code: '${response.statusCode}' for TodoNotFoundErrorResponse`,
+      );
+    }
+
+    this.statusCode = response.statusCode;
+  }
+}